Refactor and fix linter.
