+++ packages/host/src/lib/csi-controller.ts
@@ -255,7 +255,7 @@ export class CSIController extends EventEmitter {
             router.upstream("/output", this.upStreams[CommunicationChannel.OUT]);
 
 
-            router.downstream("/input", (req, res) => {
+            router.downstream("/input", (req) => {
                 const stream = this.upStreams![CommunicationChannel.IN];
                 const contentType = req.headers["content-type"];
